pyproject.toml
[tool.poetry]
name = "flask-oauth-provider-authlib-1.0.0-dev"
version = "0.1.0"
description = ""
authors = ["Pyunghyuk Yoo <62975595+purple-lion@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "1.1.2"
authlib = { git = "https://github.com/lepture/authlib.git", branch = "master" }
flask-sqlalchemy = "^2.5.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
